WIP: Move the util functions to a separate dir
